+++ docs/source/releasenotes.rst
@@ -5,12 +5,35 @@ Release notes
 `Upcoming release <https://github.com/robocorp/rpaframework/projects/3#column-16713994>`_
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-- New library **RPA.MFA** (:pr:`610`) adds support for one time passwords (OTP). Currently supports `time` 
-  and `counter` based use cases.
-- Library **RPA.Robocorp.Process** (:pr:`611`): New keywords ``List Process Run Artifacts`` and ``Get Robot Run Artifact``
 
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+16.0.0 - 31 Aug 2022
+--------------------
+
+- New library **RPA.MFA** (:pr:`610`) adds support for one time passwords (OTP).
+  Currently supports `time` and `counter` based use cases.
+- Library **RPA.Robocorp.Process** (:pr:`611`): New keywords
+  ``List Process Run Artifacts`` and ``Get Robot Run Artifact``.
+- Library **RPA.Browser.Selenium** (:issue:`494`):
+
+  - Upgraded to Selenium 4. (:pr:`602`)
+  - Using the new `webdriver-manager <https://pypi.org/project/webdriver-manager/>`_
+    for an improved download and cache of the driver. (:issue:`607`)
+  - Keyword ``Open Available Browser`` supports ``options`` parameter allowing to
+    customize the browser run. (desired capabilities got deprecated; :issue:`385`)
+
+.. warning::
+  This is a **breaking** change! The library works with the following major version
+  upgrades given any dependent package:
+
+  - ``rpaframework-aws`` **4.0.0**
+  - ``rpaframework-dialogs`` **4.0.0**
+  - ``rpaframework-google`` **6.0.0**
+  - ``rpaframework-pdf`` **5.0.0**
+  - ``rpaframework-recognition`` **5.0.0**
+  - ``rpaframework-windows`` **6.0.0**
 
 15.9.0 - 22 Aug 2022
 --------------------